+++ webapi/src/test/scala/org/knora/webapi/responders/v2/SearchResponderV2Spec.scala
@@ -260,7 +260,30 @@ class SearchResponderV2Spec extends CoreSpec() with ImplicitSender {
 
       expectMsgPF(timeout) { case response: ReadResourcesSequenceV2 =>
         compareReadResourcesSequenceV2Response(
-          expected = searchResponderV2SpecFullData.fulltextSearchForListNodeLabel,
+          expected = searchResponderV2SpecFullData.expectedResultFulltextSearchForListNodeLabel,
+          received = response
+        )
+      }
+    }
+
+    "search for list label and find sub-nodes" in {
+
+      responderManager ! FulltextSearchRequestV2(
+        searchValue = "novel",
+        offset = 0,
+        limitToProject = None,
+        limitToResourceClass = None,
+        limitToStandoffClass = None,
+        returnFiles = false,
+        targetSchema = ApiV2Complex,
+        schemaOptions = SchemaOptions.ForStandoffWithTextValues,
+        featureFactoryConfig = defaultFeatureFactoryConfig,
+        requestingUser = SharedTestDataADM.anythingUser1
+      )
+
+      expectMsgPF(timeout) { case response: ReadResourcesSequenceV2 =>
+        compareReadResourcesSequenceV2Response(
+          expected = searchResponderV2SpecFullData.expectedResultFulltextSearchForListNodeLabelWithSubnodes,
           received = response
         )
       }